pyproject.toml
[project]
name = "screen-scribe-agents"
version = "0.1.0"
description = "CrewAI + FastAPI service replacing the n8n workflows for screen-scribe-pilot."
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "pyjwt>=2.9",
    "crewai>=0.80",
    "crewai-tools>=0.14",
    "openai>=1.40",
    "pyyaml>=6.0",
    "pypdf>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.6",
]

[tool.ruff]
line-length = 100
target-version = "py311"
